move JsonRpcClient into a separate file
